refactor: Refactor after merge MEM-11
