+++ packages/client/src/assets/styles/index.scss
@@ -10,12 +10,11 @@
 
 html,
 body,
-#root,
-.App {
+#root {
   height: 100%;
 }
 
-.App {
+#root {
   display: flex;
   align-items: center;
   justify-content: center;
