+++ packages/client/src/pages/Login/constants.ts
@@ -1,11 +1,13 @@
 export const inputsData = [
   {
     id: 'login',
+    name: 'login',
     label: 'Имя пользователя',
     type: 'text',
   },
   {
     id: 'password',
+    name: 'password',
     label: 'Пароль',
     type: 'password',
   },
